+++ munguFactory/src/main/java/com/oracle/munguFactory/configuration/SecurityConfig.java
@@ -2,6 +2,7 @@ package com.oracle.munguFactory.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -10,6 +11,7 @@ import org.springframework.security.web.SecurityFilterChain;
 //IoC 빈(bean)을 등록
 @Configuration
 @EnableWebSecurity	// 필터 체인 관리 시작 어노테이션
+@EnableGlobalMethodSecurity(prePostEnabled = true, securedEnabled = true) //특정 주소 접근시 권한 및 인증
 public class SecurityConfig {
 
 	@Bean
@@ -26,7 +28,7 @@ public class SecurityConfig {
 		//Authentication 인증 (내가 사용자인가 아닌가)
 		//authorization 인가 (인증은 받았지만 권한을 확인)
 		http.authorizeRequests().antMatchers("/user/**").authenticated()
-		.antMatchers("/admin/**").access("hasRole('ROLE_MANAGER')")
+		.antMatchers("/admin/**").access("hasRole('ROLE_admin')")
 		.and().formLogin().loginPage("/loginForm").loginProcessingUrl("/login").failureUrl("/loginFail").defaultSuccessUrl("/loginSuccess");
 		// authenticated -> user/**은 인증필요-->인증만되면 들어갈 수 있음
 		//http.authorizeRequests().anyRequest().permitAll();
